refactor: move DuelManager to common
